+++ src/PrimeNumbers.java
@@ -6,11 +6,11 @@ public class PrimeNumbers {
 
     private List<Integer> findPrime() {
         int i = 0;
-        int num = 0;
+        long num = 0;
         List<Integer> primeNumbers = new ArrayList<>();
 
         for (i = 10000; i <= 99999; i++) {
-            int counter=0;
+            long counter=0;
             for(num =i; num>=1; num--) {
                 if(i%num==0) {
                     counter = counter + 1;
@@ -27,9 +27,9 @@ public class PrimeNumbers {
     public void killMyCorei3() {
         List<Integer> primeList = findPrime();
         for (int i = 0; i < primeList.size(); i++){
-            for (int number : primeList) {
-                int arg1 = primeList.get(i);
-                int arg2 = number;
+            for (long number : primeList) {
+                long arg1 = primeList.get(i);
+                long arg2 = number;
                 palindromeSearch(arg1, arg2);
             }
         }
@@ -37,7 +37,7 @@ public class PrimeNumbers {
         System.out.println("First prime is: " + palindromes.last().firstPrime);
         System.out.println("Second prime is: " + palindromes.last().secondPrine);
     }
-    private void palindromeSearch(int firstPrime, int secondPrine) {
+    private void palindromeSearch(long firstPrime, long secondPrine) {
         long palindr = firstPrime * secondPrine;
         String strResult = String.valueOf(palindr);
         if (strResult.equals(new StringBuilder(strResult).reverse().toString())) {
